[VC] Remove PerformanceTracker and add AstTraversal
Remove the PerformanceTracker utility and associated FLAG_PERFORMANCE_TRACKING logging across the codebase, eliminating per-operation perf checkpoints and Serilog perf messages. Add a new GSCode.Parser.Util.AstTraversal helper to centralize AST traversal functionality and refactor Script, Script.Helpers and diagnostics code to delegate traversal and counting to this utility. Rename and extract SPA diagnostics into SPA/ScriptDiagnosticsAnalyser and remove the now-unused NodeAnalyzer. Update ScriptManager to stop creating PerformanceTracker instances. Also add a DEBUG DefineConstants flag FLAG_MEMORY_DEBUG in the GSCode.NET.csproj. These changes simplify instrumentation, centralize AST helpers, and prepare the codebase for memory-debug builds.
