+++ server/GSCode.NET/LSP/ScriptManager.cs
@@ -258,12 +258,6 @@ public class ScriptManager
 
     private async Task<IEnumerable<Diagnostic>> ProcessEditorAsync(Uri documentUri, Script script, string content, CancellationToken cancellationToken = default)
     {
-        using var perfTracker = new PerformanceTracker("ProcessEditor", new Dictionary<string, object>
-        {
-            ["File"] = Path.GetFileName(documentUri.LocalPath),
-            ["ContentLength"] = content.Length
-        });
-
         string filePath = documentUri.LocalPath;
         var docUri = DocumentUri.From(documentUri);
         int contentHash = content.GetHashCode();
@@ -275,9 +269,7 @@ public class ScriptManager
             cached.LastParsedAt = DateTime.UtcNow;
         }
 
-        perfTracker.Checkpoint("Pre-Parse");
         await script.ParseAsync(content);
-        perfTracker.Checkpoint("Post-Parse");
 
         // Populate global symbol registry with this script's definitions (returns true if changed)
         bool symbolsChanged = PopulateSymbolRegistry(filePath, script);
@@ -290,7 +282,6 @@ public class ScriptManager
 
         // Snapshot dependencies to avoid collection modification during enumeration
         var dependencies = script.Dependencies.ToList();
-        perfTracker.AddMetadata("DependencyCount", dependencies.Count);
 
         List<Task> dependencyTasks = new();
 
@@ -300,9 +291,7 @@ public class ScriptManager
             dependencyTasks.Add(AddDependencyAsync(documentUri, dependency, script.LanguageId));
         }
 
-        perfTracker.Checkpoint("Pre-Dependencies");
         await Task.WhenAll(dependencyTasks);
-        perfTracker.Checkpoint("Post-Dependencies");
 
         // Build exported symbols
         List<IExportedSymbol> exportedSymbols = new();
@@ -319,7 +308,6 @@ public class ScriptManager
         // Merge symbols from dependencies (filtering, deduplication, conversion to relative paths)
         var (mergeFuncLocs, mergeClassLocs) = await MergeDependencySymbolsAsync(dependencies, filePath, cancellationToken);
 
-        perfTracker.Checkpoint("Pre-Analysis");
         // Merge + analyse under this script's analysis lock
         var thisDoc = DocumentUri.From(documentUri);
         await WithAnalysisLockAsync(thisDoc, async () =>
@@ -341,10 +329,8 @@ public class ScriptManager
             }
             await script.AnalyseAsync(exportedSymbols, cancellationToken);
         }, cancellationToken);
-        perfTracker.Checkpoint("Post-Analysis");
 
         var diagnostics = await script.GetDiagnosticsAsync(cancellationToken);
-        perfTracker.AddMetadata("DiagnosticCount", diagnostics.Count());
 
         return diagnostics;
     }
@@ -814,11 +800,6 @@ public class ScriptManager
 
     public async Task IndexWorkspaceAsync(string rootDirectory, CancellationToken cancellationToken = default)
     {
-        using var perfTracker = new PerformanceTracker("IndexWorkspace", new Dictionary<string, object>
-        {
-            ["RootDirectory"] = Path.GetFileName(rootDirectory)
-        });
-
         try
         {
             if (string.IsNullOrWhiteSpace(rootDirectory) || !Directory.Exists(rootDirectory))
@@ -827,16 +808,12 @@ public class ScriptManager
                 return;
             }
 
-            perfTracker.Checkpoint("Start-FileCollection");
             // Collect files first for count and deterministic iteration
             var filesList = Directory
                 .EnumerateFiles(rootDirectory, "*.*", SearchOption.AllDirectories)
                 .Where(p => p.EndsWith(".gsc", StringComparison.OrdinalIgnoreCase) ||
                             p.EndsWith(".csc", StringComparison.OrdinalIgnoreCase))
                 .ToList();
-
-            perfTracker.AddMetadata("FileCount", filesList.Count);
-            perfTracker.Checkpoint("End-FileCollection");
 
 #if DEBUG
             _logger.LogInformation("Indexing workspace under {Root}", rootDirectory);
@@ -845,12 +822,10 @@ public class ScriptManager
 #endif
 
             int maxDegree = Math.Max(1, Environment.ProcessorCount - 1);
-            perfTracker.AddMetadata("Parallelism", maxDegree);
 
             using SemaphoreSlim gate = new(maxDegree, maxDegree);
             List<Task> tasks = new();
 
-            perfTracker.Checkpoint("Start-Indexing");
             int fileIndex = 0;
             foreach (string file in filesList)
             {
@@ -873,10 +848,6 @@ public class ScriptManager
 #if DEBUG
                         _logger.LogInformation("Indexing {File}", rel);
 #endif
-                        using var fileTracker = new PerformanceTracker("IndexFile", new Dictionary<string, object>
-                        {
-                            ["File"] = rel
-                        });
 
                         await IndexFileAsync(file, cancellationToken);
 
@@ -898,7 +869,6 @@ public class ScriptManager
             }
 
             await Task.WhenAll(tasks);
-            perfTracker.Checkpoint("End-Indexing");
 
 #if DEBUG
             swAll.Stop();
@@ -915,11 +885,6 @@ public class ScriptManager
 
     private async Task IndexFileAsync(string filePath, CancellationToken cancellationToken)
     {
-        using var perfTracker = new PerformanceTracker("IndexFile-Internal", new Dictionary<string, object>
-        {
-            ["File"] = Path.GetFileName(filePath)
-        });
-
         string ext = Path.GetExtension(filePath);
         string languageId = string.Equals(ext, ".csc", StringComparison.OrdinalIgnoreCase) ? "csc" : "gsc";
 
@@ -939,13 +904,10 @@ public class ScriptManager
         // Skip if already parsed (unless it's a new file)
         if (!isNewFile && cached.Script.Parsed)
         {
-            perfTracker.AddMetadata("Skipped", true);
             return;
         }
 
-        perfTracker.Checkpoint("Pre-Parse");
         await EnsureParsedAsync(docUri, cached.Script, languageId, cancellationToken);
-        perfTracker.Checkpoint("Post-Parse");
 
         // Populate global symbol registry
         bool symbolsChanged = PopulateSymbolRegistry(filePath, cached.Script);
@@ -954,19 +916,15 @@ public class ScriptManager
 
         // Snapshot dependencies to avoid collection modification during enumeration
         var dependencies = cached.Script.Dependencies.ToList();
-        perfTracker.AddMetadata("DependencyCount", dependencies.Count);
 
-        perfTracker.Checkpoint("Pre-Dependencies");
         // Parse and register dependencies (AddDependencyAsync handles parse + registry internally)
         foreach (Uri dep in dependencies)
         {
             await AddDependencyAsync(docUri.ToUri(), dep, languageId);
         }
-        perfTracker.Checkpoint("Post-Dependencies");
 
         // Check indexing mode early to avoid unnecessary work
         var indexingMode = GSCode.Parser.Configuration.CompletionConfiguration.WorkspaceIndexingMode;
-        perfTracker.AddMetadata("IndexingMode", indexingMode.ToString());
 
         // Defensive check - IndexWorkspaceAsync should never be called with Off mode
         if (indexingMode == GSCode.Parser.Configuration.IndexingMode.Off)
@@ -976,9 +934,7 @@ public class ScriptManager
         }
 
         // Merge symbols from dependencies (filtering, deduplication, conversion to relative paths)
-        perfTracker.Checkpoint("Pre-MergeDependencies");
         var (mergeFuncLocs, mergeClassLocs) = await MergeDependencySymbolsAsync(dependencies, filePath, cancellationToken);
-        perfTracker.Checkpoint("Post-MergeDependencies");
 
         // Merge definition tables (needed for go-to-definition in both modes)
         await WithAnalysisLockAsync(docUri, async () =>
@@ -1006,13 +962,10 @@ public class ScriptManager
             // Partial indexing: signature analysis only (already done during parse)
             // Symbol registry already populated, definition tables merged
             // Skip expensive semantic analysis and exported symbols
-            perfTracker.Checkpoint("Partial-Complete");
-
         }
         else if (indexingMode == GSCode.Parser.Configuration.IndexingMode.Full)
         {
             // Full mode: build exported symbols and run semantic analysis
-            perfTracker.Checkpoint("Pre-ExportSymbols");
             List<IExportedSymbol> exportedSymbols = new();
             foreach (Uri dep in dependencies)
             {
@@ -1022,26 +975,15 @@ public class ScriptManager
                     exportedSymbols.AddRange(await depScript.Script.IssueExportedSymbolsAsync(cancellationToken));
                 }
             }
-            perfTracker.Checkpoint("Post-ExportSymbols");
 
-            perfTracker.Checkpoint("Pre-Analysis");
             await WithAnalysisLockAsync(docUri, async () =>
             {
                 // Full semantic analysis (CFG + DFA + diagnostics)
-                perfTracker.Checkpoint("Pre-AnalyseAsync");
                 await cached.Script.AnalyseAsync(exportedSymbols, cancellationToken);
-                perfTracker.Checkpoint("Post-AnalyseAsync");
             });
-            perfTracker.Checkpoint("Post-Analysis");
 
-            perfTracker.Checkpoint("Pre-PublishDiagnostics");
             // Publish diagnostics for indexed file (if LSP facade is available)
             await PublishDiagnosticsAsync(docUri, cached.Script, cancellationToken: cancellationToken);
-            perfTracker.Checkpoint("Post-PublishDiagnostics");
-
-            // Add diagnostic count for perf tracking
-            var diagnostics = await cached.Script.GetDiagnosticsAsync(cancellationToken);
-            perfTracker.AddMetadata("DiagnosticCount", diagnostics.Count);
         }
     }
