+++ server/GSCode.Parser/DFA/DataFlowAnalyser.cs
@@ -7,7 +7,6 @@ using GSCode.Parser.Lexical;
 using GSCode.Parser.SA;
 using GSCode.Parser.SPA;
 using GSCode.Parser.SPA.Logic.Components;
-using Serilog;
 
 namespace GSCode.Parser.DFA;
 
@@ -25,34 +24,10 @@ internal ref struct DataFlowAnalyser(List<Tuple<ScrFunction, ControlFlowGraph>> 
 
     public void Run()
     {
-#if FLAG_PERFORMANCE_TRACKING
-        var sw = System.Diagnostics.Stopwatch.StartNew();
-        long checkpoint1, checkpoint2;
-#endif
-
         ReachingDefinitionsAnalyser reachingDefinitionsAnalyser = new(FunctionGraphs, ClassGraphs, Sense, ExportedSymbolTable, ApiData, CurrentNamespace, KnownNamespaces, FileName, DefinitionsTable);
         reachingDefinitionsAnalyser.Run();
 
-#if FLAG_PERFORMANCE_TRACKING
-        checkpoint1 = sw.ElapsedMilliseconds;
-#endif
-
         SemanticSenseGenerator semanticSenseGenerator = new(FunctionGraphs, Sense, ExportedSymbolTable, reachingDefinitionsAnalyser);
         semanticSenseGenerator.Run();
-
-#if FLAG_PERFORMANCE_TRACKING
-        checkpoint2 = sw.ElapsedMilliseconds;
-        sw.Stop();
-        if (!string.IsNullOrEmpty(FileName))
-        {
-            Log.Debug("[PERF DETAIL] DataFlow - ReachingDefinitions: {RDA}ms, SemanticSense: {SSG}ms, Total: {Total}ms - File={File}", 
-                checkpoint1, checkpoint2 - checkpoint1, checkpoint2, FileName);
-        }
-        else
-        {
-            Log.Debug("[PERF DETAIL] DataFlow - ReachingDefinitions: {RDA}ms, SemanticSense: {SSG}ms, Total: {Total}ms", 
-                checkpoint1, checkpoint2 - checkpoint1, checkpoint2);
-        }
-#endif
     }
 }
