+++ server/GSCode.Parser/SPA/ScriptDiagnosticsAnalyser.cs
@@ -3,27 +3,48 @@ using GSCode.Parser.AST;
 using GSCode.Parser.Data;
 using GSCode.Parser.Lexical;
 using GSCode.Parser.Misc;
+using GSCode.Parser.SA;
 using GSCode.Parser.Util;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 using System.IO;
 
-namespace GSCode.Parser;
+namespace GSCode.Parser.SPA;
 
 using SymbolKindSA = GSCode.Parser.SA.SymbolKind;
 
-public partial class Script
+/// <summary>
+/// Emits basic SPA diagnostics that run after data-flow analysis in editor mode.
+/// Extracted from Script.Diagnostics.cs to follow the analyser sub-class pattern
+/// used by <see cref="GSCode.Parser.DFA.ControlFlowAnalyser"/> and
+/// <see cref="GSCode.Parser.DFA.DataFlowAnalyser"/>.
+/// </summary>
+internal ref struct ScriptDiagnosticsAnalyser(
+    ScriptNode rootNode,
+    ParserIntelliSense sense,
+    DefinitionsTable definitionsTable,
+    IReadOnlyDictionary<SymbolKey, List<Range>> references,
+    string languageId)
 {
-    // ===== SPA helpers =====
+    private ScriptNode RootNode { get; } = rootNode;
+    private ParserIntelliSense Sense { get; } = sense;
+    private DefinitionsTable DefinitionsTable { get; } = definitionsTable;
+    private IReadOnlyDictionary<SymbolKey, List<Range>> References { get; } = references;
+    private string LanguageId { get; } = languageId;
+
+    public void Run()
+    {
+        EmitUnusedParameterDiagnostics();
+        EmitUnusedUsingDiagnostics();
+        EmitUnusedVariableDiagnostics();
+        EmitAssignOnThreadDiagnostics();
+    }
 
     private void EmitUnusedParameterDiagnostics()
     {
-        if (RootNode is null) return;
-        // Traverse functions
-        foreach (var fn in EnumerateFunctions(RootNode))
+        foreach (var fn in AstTraversal.EnumerateFunctions(RootNode))
         {
-            // collect used identifiers in function body
             HashSet<string> used = new(StringComparer.OrdinalIgnoreCase);
-            CollectIdentifiers(fn.Body, used);
+            AstTraversal.CollectIdentifiers(fn.Body, used);
             foreach (var p in fn.Parameters.Parameters)
             {
                 if (p.Name is null) continue;
@@ -38,11 +59,8 @@ public partial class Script
 
     private void EmitUnusedUsingDiagnostics()
     {
-        if (RootNode is null || DefinitionsTable is null) return;
-
-        // Map referenced file paths
         HashSet<string> referencedFiles = new(StringComparer.OrdinalIgnoreCase);
-        foreach (var kv in _references)
+        foreach (var kv in References)
         {
             var key = kv.Key;
             var fLoc = DefinitionsTable.GetFunctionLocation(key.Namespace, key.Name);
@@ -51,17 +69,14 @@ public partial class Script
             if (cLoc is not null) { referencedFiles.Add(ScriptFileResolver.NormalizeFilePathForUri(cLoc.Value.FilePath)); continue; }
         }
 
-        // For each using directive, if no referenced symbol comes from that dependency file, mark unused
         foreach (var depNode in RootNode.Dependencies)
         {
-            // Build expected suffix: ..\scripts\<depNode.Path>.<LanguageId>
             string rel = depNode.Path.Replace('/', Path.DirectorySeparatorChar).Replace('\\', Path.DirectorySeparatorChar).TrimStart(Path.DirectorySeparatorChar);
             string expectedSuffix = Path.DirectorySeparatorChar + rel + "." + LanguageId;
 
             bool anyFromThisUsing = false;
             foreach (var referenced in referencedFiles)
             {
-                // referenced already normalized
                 if (referenced.Contains(expectedSuffix, StringComparison.OrdinalIgnoreCase))
                 {
                     anyFromThisUsing = true;
@@ -78,15 +93,12 @@ public partial class Script
 
     private void EmitUnusedVariableDiagnostics()
     {
-        if (RootNode is null) return;
-        foreach (var fn in EnumerateFunctions(RootNode))
+        foreach (var fn in AstTraversal.EnumerateFunctions(RootNode))
         {
-            // Count all identifier occurrences within the function body
             Dictionary<string, int> usageCounts = new(StringComparer.OrdinalIgnoreCase);
-            CollectIdentifierCounts(fn.Body, usageCounts);
+            AstTraversal.CollectIdentifierCounts(fn.Body, usageCounts);
 
-            // Single pass through function body to flag unused consts and assignments
-            foreach (var node in EnumerateChildren(fn.Body))
+            foreach (var node in AstTraversal.EnumerateChildren(fn.Body))
             {
                 if (node is ConstStmtNode cst)
                 {
@@ -103,7 +115,6 @@ public partial class Script
                 {
                     string name = id.Identifier;
                     usageCounts.TryGetValue(name, out int count);
-                    // If the only occurrence is this defining assignment (LHS), consider it unused
                     if (count <= 1)
                     {
                         Sense.AddSpaDiagnostic(id.Range, GSCErrorCodes.UnusedVariable, name);
@@ -113,46 +124,38 @@ public partial class Script
         }
     }
 
-    private static void CollectIdentifierCounts(AstNode node, Dictionary<string, int> into)
+    private void EmitAssignOnThreadDiagnostics()
     {
-        if (node is IdentifierExprNode id)
+        // Early exit: Check if there are any thread calls in the entire file first
+        bool hasAnyThreadCalls = false;
+        foreach (var node in AstTraversal.EnumerateChildren(RootNode))
         {
-            if (!into.TryGetValue(id.Identifier, out int c)) c = 0;
-            into[id.Identifier] = c + 1;
+            if (ContainsThreadCallQuickCheck(node))
+            {
+                hasAnyThreadCalls = true;
+                break;
+            }
         }
-        foreach (var child in EnumerateChildren(node))
-        {
-            CollectIdentifierCounts(child, into);
-        }
-    }
 
-    private void EmitSwitchCaseDiagnostics()
-    {
-        // Switch case diagnostics (duplicate labels, fallthrough, multiple defaults)
-        // are now handled in ControlFlowAnalyser (CFA).
-        // This method is kept as a placeholder for any future switch-specific diagnostics
-        // that don't fit in CFA.
-    }
+        if (!hasAnyThreadCalls)
+        {
+            return;
+        }
 
-    private static bool ContainsThreadCall(AstNode node)
-    {
-        if (node is PrefixExprNode pe && pe.Operation == TokenType.Thread)
+        var cache = new Dictionary<AstNode, bool>(ReferenceEqualityComparer.Instance);
+        foreach (var fn in AstTraversal.EnumerateFunctions(RootNode))
         {
-            return true;
+            foreach (var bin in AstTraversal.EnumerateBinaryExprs(fn.Body))
+            {
+                if (bin.Operation == TokenType.Assign && bin.Right is not null)
+                {
+                    if (ContainsThreadCallCached(bin.Right, cache))
+                    {
+                        Sense.AddSpaDiagnostic(bin.Range, GSCErrorCodes.AssignOnThreadedFunction);
+                    }
+                }
+            }
         }
-        if (node is CalledOnNode con)
-        {
-            // thread appears on the call side for patterns like self thread foo();
-            if (ContainsThreadCall(con.Call)) return true;
-            // also traverse 'on' to be thorough
-            if (ContainsThreadCall(con.On)) return true;
-            return false;
-        }
-        foreach (var child in EnumerateChildren(node))
-        {
-            if (ContainsThreadCall(child)) return true;
-        }
-        return false;
     }
 
     private static bool ContainsThreadCallCached(AstNode node, Dictionary<AstNode, bool> cache)
@@ -170,7 +173,7 @@ public partial class Script
         else
         {
             result = false;
-            foreach (var child in EnumerateChildren(node))
+            foreach (var child in AstTraversal.EnumerateChildren(node))
             {
                 if (ContainsThreadCallCached(child, cache)) { result = true; break; }
             }
@@ -179,53 +182,13 @@ public partial class Script
         return result;
     }
 
-    private void EmitAssignOnThreadDiagnostics()
-    {
-        if (RootNode is null) return;
-
-        // Early exit: Check if there are any thread calls in the entire file first
-        // This avoids expensive nested enumeration if there's nothing to check
-        bool hasAnyThreadCalls = false;
-        foreach (var node in EnumerateChildren(RootNode))
-        {
-            if (ContainsThreadCallQuickCheck(node))
-            {
-                hasAnyThreadCalls = true;
-                break;
-            }
-        }
-
-        if (!hasAnyThreadCalls)
-        {
-            return; // No thread calls in file, skip expensive analysis
-        }
-
-        // Proceed with full analysis only if thread calls exist
-        var cache = new Dictionary<AstNode, bool>(ReferenceEqualityComparer.Instance);
-        foreach (var fn in EnumerateFunctions(RootNode))
-        {
-            foreach (var bin in EnumerateBinaryExprs(fn.Body))
-            {
-                if (bin.Operation == TokenType.Assign && bin.Right is not null)
-                {
-                    if (ContainsThreadCallCached(bin.Right, cache))
-                    {
-                        Sense.AddSpaDiagnostic(bin.Range, GSCErrorCodes.AssignOnThreadedFunction);
-                    }
-                }
-            }
-        }
-    }
-
-    // Quick shallow check for thread token without deep recursion
     private static bool ContainsThreadCallQuickCheck(AstNode node)
     {
         if (node is PrefixExprNode pe && pe.Operation == TokenType.Thread)
         {
             return true;
         }
-        // Only check immediate children, not deep recursion
-        foreach (var child in EnumerateChildren(node))
+        foreach (var child in AstTraversal.EnumerateChildren(node))
         {
             if (child is PrefixExprNode pec && pec.Operation == TokenType.Thread)
             {
